@@ -1,9 +1,9 @@
 {
   "name": "ng-svg-icon-sprite-app",
-  "version": "1.8.1",
+  "version": "1.9.0",
   "license": "MIT",
   "author": "Jan Suwart",
-  "description": "Angular 9+ package for generating and using inline SVG icons in your project",
+  "description": "Angular 10+ package for generating and using inline SVG icons in your project",
   "bugs": {
     "url": "https://github.com/jannicz/ng-svg-icon-sprite/issues"
   },
@@ -18,8 +18,8 @@
     "build": "ng test icon-sprite --browsers=ChromeHeadless --watch=false && ng build --prod",
     "build:lib": "ng build icon-sprite --prod",
     "build:all": "npm run build:lib && npm run build",
-    "prebuild:lib": "cpr ./README.md ./projects/icon-sprite/README.md -o && cpr ./LICENSE ./projects/icon-sprite/LICENSE -o",
-    "postbuild:lib": "cpr ./projects/icon-sprite/svg-icon-sprite-anim.gif ./dist/ng-svg-icon-sprite/svg-icon-sprite-anim.gif -o",
+    "prebuild:lib": "cpr ./README.md projects/icon-sprite/README.md -o && cpr ./LICENSE projects/ng-svg-icon-sprite/LICENSE -o",
+    "postbuild:lib": "cpr projects/icon-sprite/svg-icon-sprite-anim.gif ./dist/ng-svg-icon-sprite/svg-icon-sprite-anim.gif -o",
     "test": "ng test icon-sprite --browsers=ChromeHeadless",
     "lint": "ng lint",
     "e2e": "ng e2e",
@@ -28,14 +28,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.12",
-    "@angular/common": "^9.1.12",
-    "@angular/compiler": "^9.1.12",
-    "@angular/core": "^9.1.12",
-    "@angular/forms": "^9.1.12",
-    "@angular/platform-browser": "^9.1.12",
-    "@angular/platform-browser-dynamic": "^9.1.12",
-    "@angular/router": "^9.1.12",
+    "@angular/animations": "^10.0.11",
+    "@angular/common": "^10.0.11",
+    "@angular/compiler": "^10.0.11",
+    "@angular/core": "^10.0.11",
+    "@angular/forms": "^10.0.11",
+    "@angular/platform-browser": "^10.0.11",
+    "@angular/platform-browser-dynamic": "^10.0.11",
+    "@angular/router": "^10.0.11",
     "classlist.js": "^1.1.20150312",
     "core-js": "^3.6.5",
     "rxjs": "^6.6.0",
@@ -43,11 +43,11 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.12",
-    "@angular-devkit/build-ng-packagr": "^0.1000.4",
-    "@angular/cli": "~9.1.12",
-    "@angular/compiler-cli": "^9.1.12",
-    "@angular/language-service": "^9.1.12",
+    "@angular-devkit/build-angular": "^0.1000.7",
+    "@angular-devkit/build-ng-packagr": "^0.1000.7",
+    "@angular/cli": "~10.0.7",
+    "@angular/compiler-cli": "^10.0.11",
+    "@angular/language-service": "^10.0.11",
     "@types/jasmine": "~3.5.11",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~12.12.51",
@@ -60,7 +60,7 @@
     "karma-coverage-istanbul-reporter": "~2.1.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
-    "ng-packagr": "^9.1.5",
+    "ng-packagr": "^10.0.4",
     "protractor": "~5.4.4",
     "svg2sprite-cli": "2.0.1",
     "svg4everybody": "^2.1.9",
@@ -68,7 +68,7 @@
     "tsickle": "^0.39.1",
     "tslib": "^2.0.0",
     "tslint": "^6.1.2",
-    "typescript": "^3.7.5"
+    "typescript": "^3.9.7"
   },
   "keywords": [
     "icon",
