@@ -1,9 +1,9 @@
 {
   "name": "ng-svg-icon-sprite",
-  "version": "1.8.1",
+  "version": "1.9.0",
   "license": "MIT",
   "author": "Jan Suwart",
-  "description": "Angular 9+ package for generating and using inline SVG icons in your project",
+  "description": "Angular 10+ package for generating and using inline SVG icons in your project",
   "bugs": {
     "url": "https://github.com/jannicz/ng-svg-icon-sprite/issues"
   },
@@ -12,8 +12,8 @@
     "url": "git+https://github.com/jannicz/ng-svg-icon-sprite.git"
   },
   "peerDependencies": {
-    "@angular/common": "^9.0.0",
-    "@angular/core": "^9.0.0"
+    "@angular/common": "^10.0.0",
+    "@angular/core": "^10.0.0"
   },
   "keywords": [
     "icon",
